feat(boost): introduce APOLLO_BOOST_CONFIG
