+++ packages/apollo-angular-boost/package.json
@@ -1,8 +1,7 @@
 {
   "name": "apollo-angular-boost",
   "description": "The easiest way to get started with Apollo Client",
-  "author":
-    "Kamil Kisiela <kamil.kisiela@gmail.com> (http://github.com/kamilkisiela/)",
+  "author": "Kamil Kisiela <kamil.kisiela@gmail.com> (http://github.com/kamilkisiela/)",
   "license": "MIT",
   "main": "./build/bundle.umd.js",
   "module": "./build/src/index.js",
@@ -17,9 +16,8 @@
     "build": "./node_modules/.bin/ngc -p tsconfig.build.json",
     "bundle": "rollup -c rollup.config.js",
     "test": "npm run test-only",
-    "test-only": "jest --config jest.config.js",
-    "lint":
-      "tslint -c ../../tslint.json -p tsconfig.json --type-check {src,tests}/*.ts",
+    "test-only": "npm run lint && npm run build",
+    "lint": "tslint -c ../../tslint.json -p tsconfig.json --type-check {src,tests}/*.ts",
     "postbuild": "npm run bundle",
     "deploy": "../../scripts/deploy.sh"
   },
@@ -29,33 +27,33 @@
   },
   "dependencies": {
     "apollo-cache-inmemory": "^1.2.1",
-    "apollo-angular": "^1.1.0-rc.0",
+    "apollo-angular": "^1.1.0",
     "apollo-client": "^2.3.1",
-    "apollo-angular-link-http": "^1.1.0-rc.0",
+    "apollo-angular-link-http": "^1.1.0",
     "apollo-link": "^1.2.2",
     "apollo-link-error": "^1.0.9",
     "apollo-link-state": "^0.4.1",
     "graphql-tag": "^2.9.2"
   },
   "devDependencies": {
-    "@angular/animations": "6.0.0",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/platform-server": "6.0.0",
-    "@types/graphql": "0.13.0",
-    "@types/jest": "22.2.2",
+    "@angular/animations": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/compiler-cli": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/platform-server": "6.0.2",
+    "@types/graphql": "0.13.1",
+    "@types/jest": "22.2.3",
     "@types/zen-observable": "0.5.3",
     "graphql": "0.13.2",
-    "jest": "21.2.1",
+    "jest": "22.4.4",
     "jest-zone-patch": "0.0.8",
-    "rollup": "0.57.1",
+    "rollup": "0.59.2",
     "rxjs": "6.1.0",
-    "ts-jest": "22.4.2",
+    "ts-jest": "22.4.6",
     "typescript": "2.7.2",
     "zone.js": "0.8.26"
   }
