+++ packages/apollo-angular-boost/src/api.ts
@@ -4,7 +4,9 @@ export * from 'apollo-client';
 export * from 'apollo-link';
 export * from 'apollo-cache-inmemory';
 export {Apollo} from 'apollo-angular';
-export {ApolloBoost, PresetConfig} from './ApolloBoost';
+export {PresetConfig} from './types';
+export {APOLLO_BOOST_CONFIG} from './tokens';
+export {ApolloBoost} from './ApolloBoost';
 export {ApolloBoostModule} from './ApolloBoostModule';
 export {InMemoryCache} from 'apollo-cache-inmemory';
 export {gql};
